feat: implement agent cards fetching and caching with React Query
