+++ client/webui/frontend/src/lib/api/agent-cards/keys.ts
@@ -1,0 +1,8 @@
+/**
+ * Query keys for React Query caching and invalidation
+ * Following the pattern: ['entity', ...filters/ids]
+ */
+export const agentCardKeys = {
+    all: ["agent-cards"] as const,
+    lists: () => [...agentCardKeys.all, "list"] as const,
+};
